+++ extensions/sns/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Add the new sns extension subcommand `prepare-canisters`.
+- Add the new sns extension subcommand `propose`.
 
 ## [0.1.0] - 2023-07-12